requirements.txt
torch>=2.0.0
timm>=0.6.13
pillow>=9.0.0
exifread>=3.0.0
imagehash>=4.3
opencv-python>=4.5.0
pyheif>=0.6.0